backend/requirements.txt
Django==1.11
Pillow==5.0.0
pytz==2018.3
pytesseract == 0.2.0
djangorestframework == 3.7.7